bugfix.  Now surface triangulation appears to work
Also commented-out function to fix non-optimal triangulations.
The reason is that now the surface triangulations appear to work without
this call, which means we will not be able to test it properly.  If/when
a surface arises that is clearly not delaunay triangulated, we could
uncomment this function and test it.