Working, somewhat quicker version (but can still be further optimized)
A lot of time takes place in the generation of tetrahedons in
'triangulate_domain'.  A significant part of this is expected to be in
the find_candidate_points() function, which could be sped up by
iteratively eliminating neighboring points that fall outside the current
circumscribing sphere.

Another opimization has to do with the search through the tri_flags
vector.  It has been observed that tet generation slows down over time,
and this is likely due to growth in this vector, which has to be
linearly searched through for each tet that is generated.